backend/requirements.txt
Flask
flask-cors
flask-sock
opencv-python
numpy
torch
torchvision
sounddevice
scipy
python-dotenv
gunicorn
psycopg2-binary
bcrypt
PyJWT
flask-limiter